# src/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DashboardPayload": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"data": {
						"type": "string"
					},
					"group": {
						"type": "string"
					},
					"userSub": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"data",
					"group",
					"userSub"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "cmpmp-backend",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "FindUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Finds a specified user by their id, using information gained from the express request object.\nReturns the specified user's information",
				"security": [],
				"parameters": []
			}
		},
		"/dashboard": {
			"get": {
				"operationId": "GetDashboards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Gets all dashboards and returns them",
				"security": [],
				"parameters": [
					{
						"description": "ID of a dashboard",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Name of a dashboard",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Find dashboards within a group",
						"in": "query",
						"name": "group",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Find dashboards based on a user's sub property",
						"in": "query",
						"name": "userSub",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDashboards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Creates a new dashboard using a passed in payload of dashboard information.\nIf the dashboard already exists it updates the existing dashboard (and returns a 500 error if this fails)\nIf it does not exist yet, it creates it and returns a 500 error if this fails",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DashboardPayload"
							}
						}
					}
				}
			}
		},
		"/dashboard/{id}": {
			"get": {
				"operationId": "FindDashboardById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Gets a specific dashboard using it's ID, then returns it",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteDashboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Deletes a specified dashboard using it's ID, if it can't find the specified dashboard returns a 404 error.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}